Fixed timeout bug

+++ VL53L0X.h
@@ -157,7 +157,7 @@ class VL53L0X
     uint8_t address;
     uint16_t io_timeout;
     bool did_timeout;
-    uint16_t timeout_start_ms;
+    uint32_t timeout_start_ms;
 
     uint8_t stop_variable; // read by init and used when starting measurement; is StopVariable field of VL53L0X_DevData_t structure in API
     uint32_t measurement_timing_budget_us;
@@ -176,6 +176,3 @@ class VL53L0X
 };
 
 #endif
-
-
-
